Refactor
- Moved the flask files to 'web'
- Start flask with: flask --app web run --debug
- Moved font generating files to 'font_generator'
- Updated the pytests to match
