+++ main.py
@@ -2,4 +2,4 @@ import zh
 from create_letter_key import generate_dict
 
 if __name__ == "__main__":
-    generate_dict(zh.ZH_ALL, "dict.txt")
+    generate_dict(zh.ZH_ALL, "example/dict.txt")
